requirements.txt
beautifulsoup4
lxml
pandas
numpy
matplotlib
python-telegram-bot
openpyxl
requests
asyncio
